Artificial Intelligence Based Data Governance for Chinese Electronic Health Record Analysis
Electronic health record (EHR) analysis can leverage great insights to improve the quality of human healthcare. However, the low data quality problems of missing values, inconsistency, and errors in the data set severely hinder building robust machine learning models for data analysis. In this paper, we develop a methodology of artificial intelligence (AI)-based data governance to predict the missing values or verify if the existing values are correct and what they should be when they are wrong. We demonstrate the performance of this methodology through a case study of patient gender prediction and verification. Experimental results show that the deep learning algorithm of convolutional neural network (CNN) works very well according to the testing performance measured by the quantitative metric of F1-Score, and it out performs the support vector machine (SVM) models with different vector representations for documents.